clear http endpoints (separate them); update README
clear http endpoints (separate them); update README
